# docs/swagger.yaml
definitions:
...
  models.Comment:
    properties:
      comment:
        description: Comment comment
        type: string
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      id:
        description: Model id pk
        type: integer
      post:
        $ref: '#/definitions/models.Post'
      post_id:
        description: Comment postid
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
...
  models.Permission:
    properties:
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      desc:
        description: Permission desc
        example: คำอธิบาย
        type: string
      id:
        description: Permission id
        example: 1
        type: integer
      name:
        description: Permission name
        example: create table
        type: string
      slug:
        description: Permission slug
        example: /aa
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
  models.Post:
    properties:
      comments:
        description: HasMany
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      desc:
        description: Post desc
        type: string
      id:
        description: Model id pk
        type: integer
      title:
        description: Post title
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: Post useri
        type: string
    type: object
  models.PostForCreate:
    properties:
      desc:
        description: Post desc
        type: string
      title:
        description: Post title
        type: string
    type: object
  models.PostForUpdate:
    properties:
      desc:
        description: Post desc
        type: string
      title:
        description: Post title
        type: string
    type: object
  models.Posts:
    properties:
      posts:
        items:
          $ref: '#/definitions/models.Posts'
        type: array
    type: object
  models.Role:
    properties:
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      id:
        description: Role id
        example: 1
        type: integer
      name:
        description: Role name
        example: Admin
        type: string
      permissions:
        description: Role permission
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
    type: object
  models.User:
    properties:
      created_at:
        description: Model Created At
        example: "2021-02-02 11:11:11"
        type: string
      email:
        description: User email
        example: a@email.com
        type: string
      id:
        description: Model id pk
        type: integer
      level:
        description: User level
        example: 99
        type: integer
      name:
        description: User password
        example: 'Mr. A '
        type: string
      password:
        description: User password
        example: P@ssw0rd999
        type: string
      posts:
        description: user hasMany  Post  เอา  user_id
        items:
          $ref: '#/definitions/models.Posts'
        type: array
      reftoken:
        description: User token
        example: aaaaa.bbbbbb.cccccc
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      telephone:
        description: User telephone
        example: "0816477729"
        type: string
      uid:
        description: Model UUID
        type: string
      updated_at:
        description: Model UPdated At
        example: "2021-02-02 11:11:11"
        type: string
      username:
        description: User name
        example: usera
        type: string
    type: object
...
  utils.ResponseData:
    properties:
      status:
        description: Response status
        type: integer
      success:
        default: true
        description: Response success
        example: true
        type: boolean
    type: object
info:
  contact:
    email: limweb@hotmail.com
    name: API Support
    url: http://www.servit.co.th/support
  description: '# Manageing Customer API'
  license:
    name: The MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://www.servit.co.th/
  title: ApiTest
  version: "1.0"
paths:
...
  /api/v1/posts/all:
    get:
      consumes:
      - application/json
      description: List all posts
      operationId: ListPosts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Post'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Posts
      tags:
      - posts
  /api/v1/posts/by/{id}:
    get:
      consumes:
      - application/json
      description: Get post by id
      operationId: GetPost
      parameters:
      - description: id of post to be gotten
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Basic auth: []
      summary: Get Post
      tags:
      - posts
  /api/v1/posts/del/{id}:
    get:
      consumes:
      - application/json
      description: Delete post by id
      operationId: DeletePost
      parameters:
      - description: id of post to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Delete Post
      tags:
      - posts
  /api/v1/posts/edit/{id}:
    post:
      consumes:
      - application/json
      description: Update post by id
      operationId: UpdatePost
      parameters:
      - description: id of post to be updated
        in: path
        name: id
        required: true
        type: integer
      - description: Post data to be updated
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/models.PostForUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Update Post
      tags:
      - posts
  /api/v1/posts/last:
    get:
      consumes:
      - application/json
      description: List all posts
      operationId: ListlastPosts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Post'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Posts
      tags:
      - posts
  /api/v1/posts/new:
    post:
      consumes:
      - application/json
      description: Create new post
      operationId: CreatePost
      parameters:
      - description: Post data to be created
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/models.PostForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: Create Post
      tags:
      - posts
  /api/v1/posts/vuetable:
    get:
      consumes:
      - application/json
      description: List all posts with pagination keyword filter by column Sort columns
      operationId: vuetablePost
      parameters:
      - description: number of page
        in: query
        name: page
        type: string
      - description: get number of per_page
        in: query
        name: per_page
        type: string
      - description: 'soft by cols exm: col1|asc,col2|desc'
        in: query
        name: sort
        type: string
      - description: 'search with column exm: col1|aaa,col2|bbb  by filter'
        in: query
        name: filter
        type: string
      - description: 'search by kw exm: aaa bbb ccc '
        in: query
        name: kw
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: List Posts use VueTable
      tags:
      - posts
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"